% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{diag_ext}
\alias{diag_ext}
\title{Internal function to generate diagonal matrices with possibly an offset with possibly mirrored diagonal}
\usage{
diag_ext(x, index, mirror = FALSE)
}
\arguments{
\item{x}{vector}

\item{index}{integer offset index for the diagonal (can be negative)}
}
\description{
Internal function to generate diagonal matrices with possibly an offset with possibly mirrored diagonal
}
