% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{simpson}
\alias{simpson}
\title{Internal function to calculate integration weights for Simpson's rule}
\usage{
simpson(dim)
}
\arguments{
\item{dim}{integer for the number of nodes}
}
\value{
a double vector for the integration weights
}
\description{
Internal function to calculate integration weights for Simpson's rule
}
