% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{smooth}
\alias{smooth}
\title{Internal function to smooth a vector of values using two-point average}
\usage{
smooth(v)
}
\arguments{
\item{v}{double vector to be smoother}
}
\value{
smoothed double vector
}
\description{
Internal function to smooth a vector of values using two-point average
}
