% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fredholm.R
\name{fredholm_solve}
\alias{fredholm_solve}
\title{Solve a Fredholm equation of the first and second kind}
\usage{
fredholm_solve(
  k,
  f = function(x) x,
  a = -0,
  b = 1,
  num = 41L,
  smin = 0,
  smax = 1,
  snum = 41L,
  gamma = 0.001
)
}
\arguments{
\item{k}{kernel function of two time scales}

\item{f}{left hand side function with f(a)=0}

\item{a}{lower bound of the grid for the integral approximation}

\item{b}{upper bound of the grid for the integral approximation}

\item{num}{number of points for the grid of the integral approximation}

\item{smin}{lower bound of enforcement values for equation}

\item{smax}{upper bound of enforcement values for equation}

\item{snum}{number of points for the grid for the equation}

\item{gamma}{regularization parameter}
}
\description{
Solve a Fredholm equation of the first and second kind
}
